Refactor sphinx builder from CLI
- Create new documenteer.sphinxrunner module that provides the
  run_sphinx method that actually triggers a sphinx build. This way more
  CLIs/apps can trigger Sphinx builds.
- Remove the sphinx exception handler and instead use Sphinx's own
  handle_exception method. It turns out its quite easy to create a mock
  of an argparse Namespace that is needed by handle_exception.
- Also refactor build_stack_docs() from the main CLI function. This will
  ease the transition to the new Click-based CLI.
